requirements.txt
streamlit
transformers
torch
sentence-transformers
faiss-cpu
pypdf
langchain
accelerate
